make mypy happy?
